Implement most of the core logic.
If the bulkhead profile has a definition, set the profile button's
display name (and tool tip) to the given display name and set the icon
if it is available. If there is no definition and the button does not
have a display name set, set the display name to the profile name
(better than nothing).